requirements.txt
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.15.1
idna==3.11
jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
librt==0.8.1
lightning-utilities==0.15.3
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
mypy==1.19.1
mypy-extensions==1.1.0
namex==0.1.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt-einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pillow==12.0.0
protobuf==7.34.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
simsimd==6.5.13
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.8.2
torchvision==0.22.1+cu118
triton==3.3.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
